api/requirements.txt
fastapi==0.114.1
pydantic==2.9.1
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy==2.0.34
